+++ dantro/data_loaders/load_hdf5.py
@@ -9,7 +9,7 @@ from ..base import BaseDataGroup, BaseDataContainer
 from ..containers import NumpyDataContainer
 from ..groups import OrderedDataGroup
 from ..proxy import Hdf5DataProxy
-from ..tools import fill_line
+from ..tools import fill_line, decode_bytestrings
 from ._tools import add_loader
 
 # Local constants
@@ -110,47 +110,23 @@ class Hdf5LoaderMixin:
             or DataContainer, respectively.
             """
 
-            def get_map_attr_value(attrs) -> str:
-                """As the map attribute can be a bytestring, let this function
-                retrieve it as python string.
+            def get_map_attr_val(attrs) -> str:
+                """Make sure the map attribute isn't a 1-sized array!"""
+                attr_val = attrs[map_attr] # map_attr and check in outer scope
 
-                This relies on the `map_attr` argument from the outer scope
-                """
-                attr_val = attrs[map_attr]
-
-                # Special case: numpy array of bytestring
                 if isinstance(attr_val, np.ndarray):
-                    # Convert it to a string
-                    attr_val = attr_val.tostring() 
+                    # Need be single item and already decoded
+                    attr_val = attr_val.item()
 
-                # Still might need to decode it:
-                try:
-                    return attr_val.decode('utf8')
+                return attr_val
 
-                except AttributeError:
-                    # Nope, is a string without the .decode attribute
-                    return attr_val
-
-            def decode_bytestrings(attr_val) -> str:
-                """Checks whether the given attribute value is or contains byte
-                strings and if so, decodes it to a python string.
-                """
+            def decode_attr_val(attr_val) -> str:
+                """Wrapper around decode_bytestrings"""
                 # If feature not activated, return without doing anything
                 if not self._HDF5_DECODE_ATTR_BYTESTRINGS:
                     return attr_val
 
-                # Check for data loaded as array of bytestring
-                if isinstance(attr_val, np.ndarray):
-                    if attr_val.dtype.kind in ['S', 'a']:
-                        return attr_val.astype('U')
-
-                # ... or as bytes
-                elif isinstance(attr_val, bytes):
-                    # Decode bytestring to unicode
-                    return attr_val.decode('utf8')
-
-                # Do nothing
-                return attr_val
+                return decode_bytestrings(attr_val)
 
             # Go through the elements of the source object
             for key, obj in src.items():
@@ -160,22 +136,24 @@ class Hdf5LoaderMixin:
                 if isinstance(obj, h5.Group):
                     # Need to continue recursion
                     # Extract attributes manually
-                    attrs = {k: decode_bytestrings(v)
+                    attrs = {k: decode_attr_val(v)
                              for k, v in obj.attrs.items()}
 
                     # Determine the class to use for this group
                     if enable_mapping and GroupMap and attrs.get(map_attr):
                         # Try to resolve the mapping
                         try:
-                            _GroupCls = GroupMap[get_map_attr_value(attrs)]
+                            _GroupCls = GroupMap[get_map_attr_val(attrs)]
 
                         except KeyError:
                             # Fall back to default
                             log.warning("Could not find a mapping from map "
-                                        "attribute %s='%s' to a DataGroup "
-                                        "class. Available keys: %s. Falling "
-                                        "back to default class ...",
-                                        map_attr, get_map_attr_value(attrs),
+                                        "attribute %s='%s' (originally %s) "
+                                        "to a DataGroup class. Available "
+                                        "keys: %s. Falling back to default "
+                                        "class ...",
+                                        map_attr, get_map_attr_val(attrs),
+                                        attrs[map_attr],
                                         ", ".join([k
                                                    for k in GroupMap.keys()]))
                             _GroupCls = None
@@ -219,21 +197,24 @@ class Hdf5LoaderMixin:
                         data = np.array(obj)
 
                     # Extract attributes manually
-                    attrs = {k:v for k, v in obj.attrs.items()}
+                    attrs = {k: decode_attr_val(v)
+                             for k, v in obj.attrs.items()}
 
                     # Determine the class to use for this dataset
                     if enable_mapping and DsetMap and attrs.get(map_attr):
                         # Try to resolve the mapping
                         try:
-                            _DsetCls = DsetMap[get_map_attr_value(attrs)]
+                            _DsetCls = DsetMap[get_map_attr_val(attrs)]
 
                         except KeyError:
                             # Fall back to default
                             log.warning("Could not find a mapping from map "
-                                        "attribute %s='%s' to a DataContainer "
-                                        "class. Available keys: %s. Falling "
-                                        "back to default class %s...",
-                                        map_attr, get_map_attr_value(attrs),
+                                        "attribute %s='%s' (originally %s) to "
+                                        "a DataContainer class. Available "
+                                        "keys: %s. "
+                                        "Falling back to default class %s...",
+                                        map_attr, get_map_attr_val(attrs),
+                                        attrs[map_attr],
                                         ", ".join([k for k in DsetMap.keys()]),
                                         DsetCls.__name__)
                             _DsetCls = DsetCls