+++ dantro/plot_creators/pcr_ext.py
@@ -159,9 +159,8 @@ class ExternalPlotCreator(BasePlotCreator):
             # Initialize a PlotHelper instance that will take care of figure
             # setup, invoking helper-functions and saving the figure
             hlpr = PlotHelper(out_path=out_path,
-                              enabled_helpers_defaults=plot_func.enabled_helpers,
                               helper_defaults=plot_func.helper_defaults,
-                              **(helpers if helpers else {}))
+                              update_helper_cfg=helpers)
             
             # Prepare the arguments (the data manager is added to args there)
             args, kwargs = self._prepare_plot_func_args(hlpr=hlpr,
@@ -565,9 +564,8 @@ class is_plot_func:
     """
 
     def __init__(self, *, creator_type: type=None, creator_name: str=None,
-                 use_helper: bool=True, enabled_helpers: list=None,
-                 helper_defaults: Union[dict, str]=None,
-                 **additional_attributes):
+                 use_helper: bool=True, helper_defaults: Union[dict, str]=None,
+                 add_attributes: dict=None):
         """Initialize the decorator. Note that the function to be decorated is
         not passed to this method.
         
@@ -575,11 +573,10 @@ class is_plot_func:
             creator_type (type, optional): The type of plot creator to use
             creator_name (str, optional): The name of the plot creator to use
             use_helper (bool, optional): Whether to use a PlotHelper
-            enabled_helpers (list, optional): Key strings for helpers that
-                are enabled by default
             helper_defaults (Union[dict, str], optional): Default
-                configurations for helpers in enabled_helpers
-            **additional_attributes: Additional attributes to add to the
+                configurations for helpers; these are automatically considered
+                to be enabled
+            add_attributes: Additional attributes to add to the
                 plot function
         """
         if isinstance(helper_defaults, str):
@@ -591,9 +588,8 @@ class is_plot_func:
         self.pf_attrs = dict(creator_type=creator_type,
                              creator_name=creator_name,
                              use_helper=use_helper,
-                             enabled_helpers=enabled_helpers,
                              helper_defaults=helper_defaults,
-                             **additional_attributes)
+                             **(add_attributes if add_attributes else {}))
 
     def __call__(self, func: Callable):
         """If there are decorator arguments, __call__() is only called